requirements.txt
voila
fastai2>=0.0.5
pi llow was removed
packaging
ipywidgets==7.5.1